golang-ssh: make it possible to configure connect timeout
This PR adds a way to configure (ssh.Config).Timeout, used to limix max
time spent waiting for TCP connection to enter opened state. This is
particularly important while working with firewalls which can leave
ssh.Dial hanging when destination port is not opened yet.

Additionally this PR adds NewClient ctor and Config type with proper
defaults for a conveniance.